refactor: Refactor the list item and change the styling of the work status in the avatar component.
- Move the colourStyle selector into shared types.
- Have the selected colour in the UserAvatar match the dashboard and icons.
